add package download info. fixed #63
