+++ package.json
@@ -16,6 +16,7 @@
     }
   },
   "dependencies": {
+    "humanize-number": "0.0.2",
     "gravatar": "1.0.6",
     "urllib": "0.5.5",
     "connect-markdown": "0.0.3",
