+++ view/web/layout.html
@@ -24,6 +24,9 @@
         width: 32px;
         height: 32px;
       }
+      table.downloads td.count {
+        text-align: right;
+      }
     </style>
     <script>
       $(function () {
